Migrate ActionsMenu to the SDK
