+++ frontend/packages/console-dynamic-plugin-sdk/src/api/internal-api-kubevirt.ts
@@ -516,24 +516,6 @@ export type Page<D = any> = Partial<Omit<NavPage, 'component'>> & {
 
 type NavFactory = { [name: string]: (c?: React.ComponentType<any>) => Page };
 
-export type KebabOption = {
-  hidden?: boolean;
-  label?: React.ReactNode;
-  labelKey?: string;
-  labelKind?: { [key: string]: string | string[] };
-  href?: string;
-  callback?: () => any;
-  accessReview?: AccessReviewResourceAttributes;
-  isDisabled?: boolean;
-  tooltip?: string;
-  tooltipKey?: string;
-  // a `/` separated string where each segment denotes a new sub menu entry
-  // Eg. `Menu 1/Menu 2/Menu 3`
-  path?: string;
-  pathKey?: string;
-  icon?: React.ReactNode;
-};
-
 export type FlagsObject = { [key: string]: boolean };
 
 export type WithFlagsProps = {
