+++ frontend/packages/console-dynamic-plugin-sdk/src/app/components/index.ts
@@ -4,6 +4,8 @@ export { default as PopoverStatus } from './status/PopoverStatus';
 export { default as StatusComponent } from './status/Status';
 export { default as StatusIconAndText } from './status/StatusIconAndText';
 export { default as ResourceStatus } from './utils/resource-status';
+export { ConnectedActionsMenu as ActionsMenu, ActionsMenuDropdown } from './utils/dropdown';
+export { KebabItemProps, KebabItemsProps, KebabMenuItems, kebabOptionsToMenu } from './utils/kebab';
 
 export * from './status/icons';
 export * from './status/statuses';
