+++ frontend/packages/console-shared/src/components/actions/types.ts
@@ -1,9 +1,6 @@
 import { Action, ActionGroup } from '@console/dynamic-plugin-sdk';
 
-export {
-  ActionContext,
-  ActionMenuVariant,
-} from '@console/dynamic-plugin-sdk/src/api/internal-types';
+export { ActionMenuVariant } from '@console/dynamic-plugin-sdk/src/api/internal-types';
 
 export type MenuOption = Action | GroupedMenuOption;
